Fix severe warning when using BioIK
Currently, kinematics_plugin.cpp is included in the bio_ik library and
the bio_ik_plugin library which makes them identical and results in a
severe warning by the class loader because the plugin is registered
twice. Instead, kinematics_plugin.cpp should not be included when the
library is built.